Remove unnecessary tests
We are currently skipping or ignoring tests that will never be implemented or are not applicable. I want to put an end to this practice as it means that we don’t have a way of knowing which tests really do need to implement. Since there are a lot of boiler plate tests in base classes we have a lot of test methods that can’t be removed.

This PR accomplishes test removal in two ways:

1. Add the “ignore” group to exclude tests marked as such.
2. I noticed the “testPutField” method was only being used on a handful of resources so I moved that to a trait to be included on-demand.
