+++ tests/APIv2/AbstractResourceTest.php
@@ -133,33 +133,6 @@ abstract class AbstractResourceTest extends AbstractAPIv2Test {
     }
 
     /**
-     * Test updating a field with PUT.
-     *
-     * @param string $action
-     * @param mixed $val
-     * @param string|null $col
-     * @throws \Exception if the new record already has its field set to the target value.
-     * @dataProvider providePutFields
-     */
-    public function testPutField($action, $val, $col = null) {
-        if ($col === null) {
-            $col = $action;
-        }
-        $row = $this->testPost();
-
-        $before = $this->api()->get("{$this->baseUrl}/{$row[$this->pk]}");
-        if ($before[$col] === $val) {
-            $printVal = var_export($val, true);
-            throw new \Exception("Unable to test PUT for {$this->singular} field: {$col} is already {$printVal}");
-        }
-        $urlAction = urlencode($action);
-        $this->api()->put("{$this->baseUrl}/{$row[$this->pk]}/{$urlAction}", [$col => $val]);
-        $after = $this->api()->get("{$this->baseUrl}/{$row[$this->pk]}");
-
-        $this->assertEquals($val, $after[$col]);
-    }
-
-    /**
      * Test GET /resource/<id>/edit.
      *
      * @param array|null $record A record to use for comparison.
@@ -355,14 +328,5 @@ abstract class AbstractResourceTest extends AbstractAPIv2Test {
             $r[$field] = [$field];
         }
         return $r;
-    }
-
-    /**
-     * Provide fields for PUT operations in a format that is compatible with a data provider.
-     *
-     * @return array
-     */
-    public function providePutFields() {
-        return [];
     }
 }
