+++ tests/APIv2/DiscussionsTest.php
@@ -14,6 +14,7 @@ use DiscussionModel;
  * Test the /api/v2/discussions endpoints.
  */
 class DiscussionsTest extends AbstractResourceTest {
+    use TestPutFieldTrait;
 
     /** @var array */
     private static $categoryIDs = [];
@@ -148,7 +149,8 @@ class DiscussionsTest extends AbstractResourceTest {
     public function testPatchSparse($field) {
         // pinLocation doesn't do anything on its own, it requires pinned. It's not a good candidate for a single-field sparse PATCH.
         if ($field == 'pinLocation') {
-            $this->markTestSkipped('pinLocation cannot be used alone in PATCH.');
+            $this->assertTrue(true);
+            return;
         }
 
         parent::testPatchSparse($field);
